+++ skif-service-test/skif-service-test-v1-ws/skif-service-test-v1-wsmapping/src/test/java/no/statkart/skif/skiftest/wsapi/mapping/SkifTestMappingW2DTest.java
@@ -36,9 +36,9 @@ public class SkifTestMappingW2DTest {
 
     public void testMapInt() {
         int source = 5;
-        int target = map.w2d(source);
+        Integer target = map.w2d(source);
         assertNotNull(target);
-        assertEquals(target, 5);
+        assertEquals(target.intValue(), 5);
     }
 
     public void testMapA() {
